Moved key event to its own module
